Got rid of the fake game template and add a real one in Templates.
Adding the BubbleBlower game which has you tap to blow a bubble before it shrinks away or you cause it to pop.

Made a CircleSprite to aid in circle drawing. Overriden and updated by ClickSequenceSprite so that game doesn't have duplicated code.

Fixed the update logic so it that games would get update calls, found when implementing the airLoss feature of BubbleBlower.

Updated the GlobalRuleSet with some more constants shared by all games, like CenterX/CenterY