Fixed a bug in ClickSequence when we were removing the "same entry" as a pair.
Added to the GlobalRuleSet information about titles and changed the game size to 600 by 600 to give more room for other stuff.

Added titles and help to each game.

Added a text sprite.

Added a random array index function to math.

Added code to shuffle the games list on each play which can be used when shipping and can be turned off if testing games.